requirements.txt
seashore==17.5.2
ashes==17.0.0
attrs==17.1.0
boltons==17.1.0
colorama==0.3.7
jsonpatch==1.15
jsonpointer==1.10
lithoxyl==0.4.1
marathon==0.8.11
PySocks==1.6.6
PyYAML==3.12
schema==0.6.5
requests==2.13.0
tox==2.6.0
pex==1.2.4
virtualenv==15.1.0
ptpython
gather
